責務分離リファクタ
- MetricsRegistryをinterface化しPrometheus実装とNoOp実装を分離
  メトリクス無効時もnon-nullになり呼び出し側のnullチェックを削除
- ITTSRuntimeUseにgetTTSCountRecorder/getMetricsRegistryのdefaultを追加
  VoiceAudioSchedulerのITTSRuntime.getInstance()プルを既存パターンに統一
- ITTSRuntime#initMetricsを生成と起動に限定し、シャットダウンフック登録を
  registerShutdownHooksに分離
