+++ core/src/main/java/dev/felnull/itts/core/ITTSRuntimeUse.java
@@ -5,6 +5,8 @@ import dev.felnull.itts.core.cache.CacheManager;
 import dev.felnull.itts.core.config.ConfigManager;
 import dev.felnull.itts.core.dict.DictionaryManager;
 import dev.felnull.itts.core.discord.Bot;
+import dev.felnull.itts.core.metrics.MetricsRegistry;
+import dev.felnull.itts.core.tts.TTSCountRecorder;
 import dev.felnull.itts.core.tts.TTSManager;
 import dev.felnull.itts.core.voice.VoiceManager;
 import org.apache.logging.log4j.Logger;
@@ -82,5 +84,24 @@ public interface ITTSRuntimeUse {
      */
     default ITTSNetworkManager getNetworkManager() {
         return getITTSRuntime().getNetworkManager();
+    }
+
+    /**
+     * 読み上げ文字数のレコーダーを取得
+     *
+     * @return TTSカウントレコーダー
+     */
+    default TTSCountRecorder getTTSCountRecorder() {
+        return getITTSRuntime().getTTSCountRecorder();
+    }
+
+    /**
+     * メトリクスレジストリを取得
+     * 公開無効時もNoOp実装が返るためnullにはならない
+     *
+     * @return メトリクスレジストリ
+     */
+    default MetricsRegistry getMetricsRegistry() {
+        return getITTSRuntime().getMetricsRegistry();
     }
 }
